Updated examples and consolidated files
