+++ torchtune/modules/model_fusion/__init__.py
@@ -4,9 +4,7 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-from ._fusion_embed import FusionEmbedding
-from ._fusion_layer import FusionLayer
-from ._fusion_models import DeepFusionModel
+from ._fusion import DeepFusionModel, FusionEmbedding, FusionLayer
 from ._fusion_utils import register_fusion_module
 
 __all__ = [
