+++ torchtune/modules/model_fusion/_fusion_utils.py
@@ -15,6 +15,16 @@ def register_fusion_module(module: nn.Module):
     marks all of the Modules parameters as fusion params.
     This can be used for a layer or an entire model that is
     added to combine two or more pretrained models.
+
+    For example, you might want to add a projection head
+    head onto an encoder to learn a projection from the
+    pre-trained encodings to the decoder's embedding space. This
+    is typical with both Deep Fusion and Early Fusion models.
+
+    Example::
+        >>> projection_head = FeedForward(...)
+        >>> register_fusion_module(projection_head))
+        >>> encoder = nn.Sequential(clip_vit_224(), projection_head)
     """
 
     def fusion_params(self) -> List[str]:
